+++ Software/src/LedDeviceVirtual.hpp
@@ -46,7 +46,7 @@ public slots:
     void setGamma(double value);
     void setBrightness(int value);
     void requestFirmwareVersion();
-    int maxLedsCount() { return 255;}
+    size_t maxLedsCount() { return 255;}
 
 
 private:
